backend/requirements.txt
asgiref==3.9.2
attrs==25.3.0
click==8.3.0
colorama==0.4.6
coverage==7.4.0
django==4.2.24
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
drf-spectacular==0.28.0
global-land-mask==1.0.0
gunicorn==22.0.0
inflection==0.5.1
iniconfig==2.1.0
jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markupsafe==3.0.3
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.2.250926
pip==25.2
pluggy==1.6.0
psycopg2-binary==2.9.9
pydeck==0.9.1
pytest==7.4.4
pytest-django==4.7.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing-extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
uv==0.8.22
wait-for-it==2.2.2
wheel==0.45.1
